+++ src/editor/config.js
@@ -47,6 +47,7 @@ export const LOWERCASE_TAGS = generateKeyHash([
 
 export const CLASS_OR_ID = genUpper2LowerKeyHash([
   'mousetrap',
+  'AG_THEME_ID',
   'AG_GRAY',
   'AG_HIDE',
   'AG_WARN',